+++ src/backend/storage/buffer/buf_init.c
@@ -172,6 +172,7 @@ InitBufferPool(void)
 
 			pg_atomic_init_u32(&buf->state, 0);
 			pg_atomic_init_u32(&buf->polar_redo_state, 0);
+			pg_atomic_init_u64(&buf->polar_outdate_lsn, InvalidXLogRecPtr);
 			buf->wait_backend_pgprocno = INVALID_PGPROCNO;
 
 			buf->buf_id = i;